Modified tests
